Del task hook  fetches depending on session

+++ src/hooks/useHandleDeleteTask.ts
@@ -4,23 +4,26 @@ import { useAppDispatch } from "./useAppDispatch"
 import { deleteTask } from "features/tasksSlice"
 import { setSnackbar, setSnackbarError } from "features/appSlice"
 //Actions to do on delete task
-export default function useHandleDeleteTask() { 
+export default function useHandleDeleteTask() {
     const userid = useAppSelector(state => state.auth.userid)
+    const hasSession = useAppSelector(state => state.auth.hasSession)
     const dispatch = useAppDispatch()
 
     const [deleteTaskMut, deleteResult] = useDeleteTaskMutation()
 
-    async function handleDeleteTask(taskId: number) { 
-        try{
-            const response = deleteTaskMut({
-                userId: userid!,
-                taskId: taskId
-            }).unwrap() 
-            console.log(response)
+    async function handleDeleteTask(taskId: number) {
+        try {
+            if (hasSession && (userid !== null)) {
+                const response = deleteTaskMut({
+                    userId: userid!,
+                    taskId: taskId
+                }).unwrap()
+                console.log(response)
+            }
             dispatch(deleteTask(taskId));
             dispatch(setSnackbar("Task deleted!"))
         }
-        catch(error){ 
+        catch (error) {
             console.log(error)
             dispatch(setSnackbarError("Error deleting task"))
         }
